Services return the HTTP call's status
In order to properly record the data for ServiceEvents in codeclimate
proper, we need to have the calls to the external services return
something useful.

The way this works is by moving all of the actual HTTP POSTing into a
dedicated method that wraps the response. Failures are handled by the
`WithErrorHandling` invocation middleware as normal. A successful
request will be parsed for standard things like status code, the params
we sent, the url we hit, and whether it worked or not, and a message.
Failures will have a message explaining the error. `receive_test`
methods will include a user-friendly error message. Any other error will
forego HTTP information. This is all tested and common to each service, so
they don't need their own specific tests for it.

A successful request can be modified by adding extra attributes into the
hash via a block. The response body is passed to the block to be parsed
inside it (since we get both JSON and XML from various services).
Relevant values can be extracted in this block. Parse errors will
trigger the generic error handling and return a "this failed" message to
the caller. The uniquely-added attributes are tested on a per-service
basis.

Some services make multiple requests. The first one to fail is what we
return an error message for. If none fail, it's all good.