Display the error message when we have one
Both Asana and Github provide a nicer error message when something goes
wrong. If we get that response body, use it by supplying it as a
supplementary error message. If we have this message, prefer it to the
"real" exception's message. We must still send the message we log, so
that it's saved along with the event's data.