Update GitHub status description for skipped builds
https://trello.com/c/q3ByzTqY

When analysis for a commit is skipped because a newer commit was pushed
to the branch, we want to set the status to "success" so that it isn't
stuck in a yellow "pending" state, but we also want to indicate that
analysis didn't run on the skipped commit. So we differentiate the
skipped case with a different status message. Also, never add a comment
for a commit that we have skipped, regardless of the config settings.