Ensure any result that's nil result becomes useful
Add a middleware that should be placed at the bottom of the Invocation
stack. This middleware ensures that any time the invocation stack is
called, there is a return value. If the stack comes back nil, then a
message is put in its place implying an error. While this might not
technically be the case that the service errored, a lack of return value
should be considered an error and looked in to.