Return hash instead of nil when no changes in snapshot
https://trello.com/c/FVI4JXBI

Before this change, attempting to send a snapshot to Slack when there
were no alert or improved constants would return nil, which is an
unexpected value in the app. This change checks to see if those changes
exist before attempting to send a snapshot, and if they don't, returns
a hash with ignored: true instead of nil.