Follow redirects for GH issue creation (#97)
* Follow redirects for GH issue creation

I tracked this down related to a CX escalation. Issue creation URLs for
at least some repos are currently returning 307 statuses: [GH's docs][1]
make it fairly clear that we should not be changing the URL we start
with, but we should be following these. I was locally able to confirm
this fixed issue creation for an affected repo.

[1]: https://developer.github.com/v3/#http-redirects

* bump ruby version: needed for newer activesupport we're using