Add support for welcome comments on GH PRs
This change also introduces a refactoring of how we generate service
responses.

I needed this service to include responses with metadata for multiple
network requests, and making that possible required this refactoring.

Also removed the code related to following redirects because we don't
currently use it, and it was simpler to do the refactor that way.