Transform test suite to RSpec (#116)
We noticed we weren't getting correct test coverage results on this repo. Digging into it, this test suite was still on Test::Unit, and simplecov appears to not work properly with that anymore.

I've converted the suite to RSpec, which brings it more in line with our other projects. Locally running SimpleCov also shows that it appears to fix the inaccurate coverage results as well.

This conversion was done mostly with a bunch of local sed scripting & manually fixing cases I couldn't handle automatically, so there's probably still some messy bits. I personally don't think we need to get this perfect in every little bit of the test suite. Tests pass with no warnings, so I think it's pretty good.