Allow rolling out PR status reporting
Adds new config values to allow reporting PR statuses for only specific
users and/or a percentage of users.

The `author_username` key being used in the payload isn't actually sent
right now: I need to evaluate if there's an existing key we send that is
usable, or if we need to thread this data through on the app side.