SafeWebhook
This implements a check that outgoing webhooks aren't making connections
to internal services. It does this by resolving the webhook's host and
checking it against internal subnet masks.

Crucially, it must also adjust the default `Resolv` behavior to ensure
that the eventual connection gets the same IP address as the one we've
verified as non-internal.

All of this is override-able via ENV, since an on-prem installation will
want to intentionally target internal services with their webhooks.