Support authorizing asana requests via tokens
Asana is deprecating their API Key access, as reported by a helpful
user.

https://asana.com/developers/documentation/getting-started/auth#api-key

We confirmed with their support team that they're planning to drop
support for API Key authorization in "early 2017".

This PR introduces support for authorizing with personal access tokens,
and attempts to nudge users in that direction gently. I think we'll need
to make a plan about how to migrate existing repos over.