Respect proxy exceptions
Our usage of Faraday in this library is...unusual.

Faraday does generally respect the proxy & proxy exception settings when
you pass the URL directly to a `conn.post` method or anything like that,
but since we set the URL within this block, that's after Faraday's logic
has decided what the proxy should be, and since at that point Faraday
didn't know what URL we were requesting, it used the default proxy
details.